PDBSessionMiddleware: Non-API sessions for CSRF
Because we use CsrfViewMiddleware, we need to include CSRF tokens with
POST requests by default. Because we use CSRF_USE_SESSIONS, we need
to be able to create a session in order to save the CSRF token.

Currently a "CSRF cookie not set" error is returned from some endpoints.
I reproduced on prod by creating a new account, then deleting my session
cookie, and then opening the email confirmation link and trying to
submit it. Even if we were to consider dropping CSRF support, it's not
particularly easy to do so for views which are in external modules.

Instead of allowlisting specific endpoints which will be allowed to
create sessions, create a denylist of endpoints where sessions should
not be created (i.e. API endpoints).

This should allow us to gain the benefit of the reduced session overhead
for API calls, while also allowing us to use CSRF protection for
anonymous users. I would like to use this to add a language selection
drop-down which will POST to the Django i18n view's set_language
endpoint.