Site Header: Add language selection drop-down
Anonymous users can use this to set the language of any page without
having to create a profile, and the setting will be persisted in their
browser with a cookie.

Logged-in users can use this to set the language cookie as well, but it
will not update the setting in their profile.